pyproject.toml
[tool.poetry]
name = "sub-maker"
version = "0.1.0"
description = ""
authors = ["Martin <ma.barton@seznam.cz>"]
readme = "README.md"
packages = [{include = "sub_maker"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.1"
ffmpeg-python = "^0.2.0"
codetiming = "^1.4.0"
srtranslator = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
